LG4 Rename ScopeCondition module file
