+++ model/scopes_conditions.py
